requirements.txt
asgiref==3.9.1
coverage==7.10.6
Django==5.2.6
django-recurrence==1.11.1
djangorestframework==3.16.1
drf-nested-routers==0.94.2
numpy==2.3.2
pandas==2.3.2
psycopg2-binary==2.9.10
pyaml==25.7.0
python-dateutil==2.9.0.post0
python-decouple==3.3
pytz==2025.2
PyYAML==6.0.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1
